Update non-autowirable-args.md - Round #1
Is it spelled "autowirable" or "autowireable"? I thought it was the latter, but I may be wrong. If I'm wrong, the spelling "autowireable" with an "e" appears twice in the script (title and last paragraph) and that will need to be changed. If I'm right, the file may need to be renamed for consistency.