Update autoconfiguration.md - Round #1
This was a LONG one and I did quite a bit of rephrasing to help the flow. Three lines to pay attention to in particular are 25, 41, and 42. I was having some difficulty understanding exactly what was happening so my rewording in those sections is probably the least accurate. I hope I got somewhat close, at least.